requirements.txt
altgraph==0.17.2
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
Backtesting==0.3.2
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
bokeh==2.4.2
boltons==21.0.0
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1625831756778/work
chardet @ file:///C:/ci/chardet_1607706937985/work
charset-normalizer==2.0.12
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
conda==4.11.0
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262410900/work
configparser==5.0.2
cryptography @ file:///C:/ci/cryptography_1616769504165/work
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.3
docopt==0.6.2
entrypoints==0.4
environs==9.5.0
et-xmlfile==1.1.0
executing==0.8.2
feedparser==6.0.8
ffn==0.3.6
filelock==3.1.0
finlab==0.1.40.dev1
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
hanziconv==0.3.2
haohaninfo==0.2.21
helium==3.0.8
html5lib==1.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
influxdb-client==1.26.0
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kafka-python==2.0.2
kiwisolver==1.3.2
line-bot-sdk==1.20.0
lxml==4.6.3
lz4==4.0.0
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
menuinst==1.4.16
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.3
multipledispatch==0.6.0
multitasking==0.0.9
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy @ file:///D:/bld/numpy_1636145505783/work
opencc-python-reimplemented==0.1.6
openpyxl==3.0.9
opstrat==0.1.7
packaging==21.3
pandas==1.3.4
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pefile==2021.9.3
pendulum==2.1.2
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
plotly==5.6.0
prettytable==2.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.8.0
pure-eval==0.2.2
pyarrow==7.0.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.11.2
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
pykalman==0.9.5
pymongo==3.12.0
pynput==1.7.6
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.2
pyrsistent==0.18.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.7
pytz==2021.1
pytzdata==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
redis==3.5.3
requests==2.27.1
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
Rx==3.2.0
scikit-learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
simplejson==3.17.5
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
soupsieve==2.2.1
sourcedefender==7.1.67
stack-data==0.2.0
statsmodels==0.12.2
sympy==1.8
TA-Lib @ file:///D:/bld/ta-lib_1637267139956/work
tabulate==0.8.9
tenacity==8.0.1
terminado==0.13.1
testpath==0.5.0
TgCrypto==1.2.3
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.1.1
tushare==1.2.64
typing_extensions==4.0.1
tzdata==2021.1
tzlocal==3.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
virtualenv==20.8.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
wget==3.2
widgetsnbextension==3.5.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
xlwings==0.24.9
yfinance==0.1.63
zmq==0.0.0
